requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
tensorflow==2.18.0
Pillow==10.0.0
numpy>=1.24.3
google-generativeai==0.3.1
firebase-admin==6.2.0
pyrebase4==4.7.1
requests<=2.29.0
pytz==2023.3.post1
reportlab==4.1.0
pytest==7.4.2
gcloud==0.18.3
python-jwt==4.0.0
cryptography==41.0.4
Jinja2==3.1.2
onnxruntime==1.16.3
transformers==4.35.2
opencv-python==4.8.1.78
matplotlib==3.8.2